testing to split coroutines and reactor
